feat: button, sidebar style 적용
